Edits to Gp to save observation/validation set
Also small changes to the code so that the entire dataset is saved, not just the datetime column. I created a second column with the datatime in minutes (instead of hours/seconds) and did some renaming so that the columns are clear. NOTE: the data cleaning process resulted in slightly different number of rows for me than for Manli